requirements.txt
Django
python-telegram-bot
aiohttp
playwright
reportlab
celery
redis